backend/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
kafka-python==2.0.2
python-json-logger==2.0.7
python-dotenv==1.0.0
